move read_taxa to amplicon
